sendAccountActivationEmail in separate file
